+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/MandateHandler.kt
@@ -54,7 +54,8 @@ internal class MandateHandler(
                 selection = viewModel.selection,
                 merchantDisplayName = viewModel.config.merchantDisplayName,
                 isVerticalModeProvider = {
-                    viewModel.paymentMethodMetadata.value?.paymentMethodOrientation == PaymentMethodOrientation.Vertical
+                    viewModel.paymentMethodMetadata.value?.paymentMethodOrientation() ==
+                        PaymentMethodOrientation.Vertical
                 },
                 isSetupFlowProvider = { viewModel.paymentMethodMetadata.value?.stripeIntent is SetupIntent },
             )